+++ packages/eslint-plugin/rules/__tests__/use-recommended-components.js
@@ -131,7 +131,8 @@ ruleTester.run( 'use-recommended-components', rule, {
 			code: "import { BaseControl } from '@wordpress/components';",
 			errors: [
 				{
-					message: 'Use `Field` from `@wordpress/ui` instead.',
+					message:
+						'Use `Field` from `@wordpress/ui` instead. For a purely visual label, use `Field.VisualLabel`.',
 				},
 			],
 		},