+++ packages/ui/src/form/primitives/field/index.ts
@@ -5,15 +5,5 @@ import { Description } from './description';
 import { Details } from './details';
 import { Control } from './control';
 import { VisualLabel } from './visual-label';
-import { VisualDescription } from './visual-description';
 
-export {
-	Root,
-	Item,
-	Label,
-	Description,
-	Details,
-	Control,
-	VisualLabel,
-	VisualDescription,
-};
+export { Root, Item, Label, Description, Details, Control, VisualLabel };